Helper: moved InitiallyDefinition and InitializerDefinition
