+++ src/transform/AstDumpSourcePass.cpp
@@ -113,7 +113,6 @@ void AstDumpSourceVisitor::visit( InitDefinition& node )
 
 void AstDumpSourceVisitor::visit( InitializerDefinition& node )
 {
-    node.attributes()->accept( *this );
     node.leftBraceToken()->accept( *this );
     node.arguments()->accept( *this );
     node.rightBraceToken()->accept( *this );
